+++ solr/core/src/java/org/apache/solr/util/hll/HLL.java
@@ -203,7 +203,7 @@ public class HLL implements Cloneable {
             // TODO improve this cutoff to include the cost overhead of Java
             //      members/objects
             final int largestPow2LessThanCutoff =
-                    (int)NumberUtil.log2((this.m * this.regwidth) / this.shortWordLength);
+                    (int)NumberUtil.log2((this.m * this.regwidth) / (double) this.shortWordLength);
             this.sparseThreshold = (1 << largestPow2LessThanCutoff);
         }
